Build with sample-go-multiarch
Switch from using Shipwright to a sample go application. This app will
be used in a subsequent demo for mult-arch builds on Kubernetes. Since
this sample uses Helm to deploy, we can't use `ko resolve` easily. The
sample instead has Makefile targets which simplify the build and deploy
on a cluster.

Signed-off-by: Adam Kaplan <adam@adambkaplan.com>